Move fixtures into the backend module
